requirements.txt
selenium>=4.0.0
webdriver-manager>=3.5.2
openai>=0.27.0
pyyaml>=6.0
pandas>=1.3.0
tqdm>=4.62.0
jupyter>=1.0.0
pytest>=6.2.5
